Support Contents Menu Launch Action
* move scripture display to /text route
* fix firstLaunch so that it can be handled by start page
  and text page
* handle query string parameters in start page
* use navigate function to add query string ref to history
